Revert "only output stats relevant to the current m-line direction"
It's wrong to not count m-lines that are not sending or receiving
because what we want to count is also the total peer traffic.

If we exclude a track on a change of direction the byte counts
would decrease as if there was a negative traffic, which does
not make sense.

Note that this would happen anyway in case we start Drop-ing
senders/receivers sources, which is where we store stats.

This reverts commit ff59031b29aaa9ba757167570def7226096aa92a.