Continuously updated QueueState in PacketizerBuffer
This splits QueueState into two halves, QueueState/QueueSnapshot and
the snapshot is used on the PacketizingBuffer level and QueueState on
the MediaInner level.  This is to avoid needing all knowledge about
what kind of data we're handling inside the PacketizingBuffer.

The more important change is adding an internal TotalQueue struct in
PacketizingBuffer that continuously keeps track of the total queue
duration in the buffer. This way we can avoid an expensive
recalculation on every MediaInner::queue_state() call.