Restrict periodic BWE adjustments
We never stopped periodically adjusting BWE based on the current
trendline hypothesis, regardless of how old the hypothesis was. This
meant that if we had a hypothesis of underuse and stopped receiving TWCC
we would indefinitely increase the bandwidth estimate.

With this change we'll only consider a trendline hypothesis valid for a
while after the TWCC report that updated it.