buffer_tx: add configurable duration threshold
Allows the buffer to free packets that are older than a configured threshold.
This refers to RTP time and acts within the boundaries of the buffer size
without changing the size of the buffer itself.

When packets are evicted the Packetized::data Vec<u8> is dropped,
therefore freeing up memory. This will be more efficient as we reduce
the overhead of the Packetized structure itself.

Having a larger size and limiting by time allows to more efficiently accomodate
spikes due to big keyframes (i.e. think of screenshares).