Simplify TWCC egress loss
Considering the RTT and likelihood that the TWCC report has been sent
and received is a tad more robust but it can cause spurious incorrect
values instead. This simplified version simply ignores packets that we
haven't received a TWCC report for. If a whole TWCC report is lost we'll
have less data, but as long as the loss is somewhat uniform this
shouldn't matter much.