Merge stats for a given (mid, rid) pair
A given (mid, rid) pair can use multiple SSRCs. Not only is there an RTX
SSRC and a main SSRC, which was already handled prior to this commit,
but during the lifetime of an RTP Session new SSRCs (RTX and main) can be
created for the same (mid, rid) pair. In particular, Firefox does this
when a transceiver transitions from send to inactive, and back to send.
When media is resumed the original SSRCs are not used, instead new SSRCs
are created.