Prefer str0m's locked pts over offer suggestions
When receiving an offer with a `sendonly` m-line don't remap Pts if they
differ from str0m's locked values. The values from the offer are what
the offerer would like to send, but ultimately the answer dictates this.
If str0m has already establish a pt for the codecs in question reply
with those.