Remove DtlsCert::new() on RtcConfig::new()
Creating new dtlscert takes time and is expensive,
in a server scenario one might want to reuse the certs.

This change allow the user to create it's own cert and
then clone it for new instances instead of str0m creating
a new one on every configure.

If a user have not created a cert then the build will
generate one.