Respect `ssrc-group` simulcast signalling
Firefox has bugs with dynamically establishing the mappings of SSRCs for
simulcast using the rid and repaired-rid RTP header, however it does
send the `ssrc-group` correctly in SDPs. To workaround these bugs we
need to give preference to the mapping signalled via ssrc-group, if
enabled.

When SSRCs are explicitly signalled in SDP we disable the rid and
repaired-rid headers and rely on these signalled SSRCs.