MediaTime: use a non-zero u32 (newtype Frequency) as denominator
Ensure that we can never have a delayed divide by zero panic due to
carrying a zero denominator around. This removes a denial of service
vulnerability from SDP parsing and makes the MediaTime type safer for
API users. The change from 64-bit to 32-bit is justified by the
observation that it is unlikely applications will need to represent
times with precision smaller than quarters of nanoseconds.