fix(ice): do not disconnect whilst we still check new candidates (#489)
In case the active candidate pair gets invalidated, str0m currently instantly emits a `Disconnected` event. For example, if a candidate pair with a peer-reflexive candidate gets nominated and later replaced. This results in a disconnected event despite us not actually being disconnected (a peer-reflexive candidate that gets replaced is after all the exact same socket).

To avoid this false-positive disconnection, we transition back to `Checking` in case we are still evaluating other candidates.