Wincrypto Implementation for Str0m (#589)
Implement new Str0m feature 'wincrypto'.

When enabled, this cause Str0m to use Windows Cryptographic APIs for crypto.
This feature cannot be used in conjunction with 'sha1' not 'openssl' features, as it provides the same functionality.

Windows APIs are accessed via the windows-rs crate. These calls are `unsafe`, so in order to isolate them from
`safe` str0m code, all `unsafe` code is in a separate `str0m_wincrypto` crate. The code in the core `str0m/crypto/wincrypto` is the glue code between Str0m and this str0m_wincrypto crate.

The `str0m_wincrypto` crate is not intended to be a generic crypto crate, and is very much tailored to match str0m's crypto APIs.