feat(ice): go to `Checking` from `Disconnected` on new candidates (#676)
It is recommened to not immediately act on `IceConnectionState::Disconnected` as the connection may be restored once new candidates get added. Currently though, `str0m` will directly transition from `Disconnected` to `Connected` and completely skip the `Checking` state.

To be consistent with other state transition, we now transition from `Disconnected` to `Checking` first before moving on to `Connected` (if any of the new candidates turn out to be viable).