chore(ice): consistent TRACE logs in candidate pair (#682)
Sometimes, it is useful to just see logs for which candidate pairs are sending and receiving binding requests. To achieve this, we extend the existing TRACE logs to also log the full candidate pair whenever we receive a request and send a request.

Combined with the existing logs, this makes a log filter of `str0m::ice_::pair=trace` quite useful to look at.